# Make Enumerable.First/Last/GetElement faster on various enumerables