--- a/src/libraries/System.Linq/src/System/Linq/Concat.SpeedOpt.cs
+++ b/src/libraries/System.Linq/src/System/Linq/Concat.SpeedOpt.cs
@@ -89,6 +89,65 @@ namespace System.Linq
                     return result;
                 }
             }
+
+            public override TSource? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    foreach (IEnumerable<TSource> source in (ReadOnlySpan<IEnumerable<TSource>>)[_first, _second])
+                    {
+                        if (TryGetNonEnumeratedCount(source, out int count))
+                        {
+                            if (index < count)
+                            {
+                                found = true;
+                                return source.ElementAt(index);
+                            }
+
+                            index -= count;
+                        }
+                        else
+                        {
+                            using IEnumerator<TSource> e = source.GetEnumerator();
+                            while (e.MoveNext())
+                            {
+                                if (index == 0)
+                                {
+                                    found = true;
+                                    return e.Current;
+                                }
+
+                                index--;
+                            }
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public override TSource? TryGetFirst(out bool found)
+            {
+                TSource? result = _first.TryGetFirst(out found);
+                if (!found)
+                {
+                    result = _second.TryGetFirst(out found);
+                }
+
+                return result;
+            }
+
+            public override TSource? TryGetLast(out bool found)
+            {
+                TSource? result = _second.TryGetLast(out found);
+                if (!found)
+                {
+                    result = _first.TryGetLast(out found);
+                }
+
+                return result;
+            }
         }
 
         private sealed partial class ConcatNIterator<TSource> : ConcatIterator<TSource>
@@ -210,9 +269,80 @@ namespace System.Linq
 
                 return array;
             }
+
+            public override TSource? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    IEnumerable<TSource>? source;
+                    for (int i = 0; (source = GetEnumerable(i)) is not null; i++)
+                    {
+                        if (TryGetNonEnumeratedCount(source, out int count))
+                        {
+                            if (index < count)
+                            {
+                                found = true;
+                                return source.ElementAt(index);
+                            }
+
+                            index -= count;
+                        }
+                        else
+                        {
+                            using IEnumerator<TSource> e = source.GetEnumerator();
+                            while (e.MoveNext())
+                            {
+                                if (index == 0)
+                                {
+                                    found = true;
+                                    return e.Current;
+                                }
+
+                                index--;
+                            }
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public override TSource? TryGetFirst(out bool found)
+            {
+                IEnumerable<TSource>? source;
+                for (int i = 0; (source = GetEnumerable(i)) is not null; i++)
+                {
+                    TSource? result = source.TryGetFirst(out found);
+                    if (found)
+                    {
+                        return result;
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public override TSource? TryGetLast(out bool found)
+            {
+                ConcatNIterator<TSource>? node = this;
+                do
+                {
+                    TSource? result = node._head.TryGetLast(out found);
+                    if (found)
+                    {
+                        return result;
+                    }
+                }
+                while ((node = node!.PreviousN) is not null);
+
+                found = false;
+                return default;
+            }
         }
 
-        private abstract partial class ConcatIterator<TSource> : IIListProvider<TSource>
+        private abstract partial class ConcatIterator<TSource> : IPartition<TSource>
         {
             public abstract int GetCount(bool onlyIfCheap);
 
@@ -236,6 +366,17 @@ namespace System.Linq
 
                 return list;
             }
+
+            public abstract TSource? TryGetElementAt(int index, out bool found);
+
+            public abstract TSource? TryGetFirst(out bool found);
+
+            public abstract TSource? TryGetLast(out bool found);
+
+            public IPartition<TSource>? Skip(int count) => new EnumerablePartition<TSource>(this, count, -1);
+
+            public IPartition<TSource>? Take(int count) => new EnumerablePartition<TSource>(this, 0, count - 1);
+
         }
     }
 }