--- a/src/libraries/System.Linq/src/System/Linq/Where.SpeedOpt.cs
+++ b/src/libraries/System.Linq/src/System/Linq/Where.SpeedOpt.cs
@@ -8,7 +8,7 @@ namespace System.Linq
 {
     public static partial class Enumerable
     {
-        private sealed partial class WhereEnumerableIterator<TSource> : IIListProvider<TSource>
+        private sealed partial class WhereEnumerableIterator<TSource> : IPartition<TSource>
         {
             public int GetCount(bool onlyIfCheap)
             {
@@ -68,9 +68,90 @@ namespace System.Linq
 
                 return list;
             }
+
+            public TSource? TryGetFirst(out bool found)
+            {
+                Func<TSource, bool> predicate = _predicate;
+
+                foreach (TSource item in _source)
+                {
+                    if (predicate(item))
+                    {
+                        found = true;
+                        return item;
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TSource? TryGetLast(out bool found)
+            {
+                using IEnumerator<TSource> e = _source.GetEnumerator();
+
+                if (e.MoveNext())
+                {
+                    Func<TSource, bool> predicate = _predicate;
+                    TSource? last = default;
+                    do
+                    {
+                        TSource current = e.Current;
+                        if (predicate(current))
+                        {
+                            last = current;
+                            found = true;
+
+                            while (e.MoveNext())
+                            {
+                                current = e.Current;
+                                if (predicate(current))
+                                {
+                                    last = current;
+                                }
+                            }
+
+                            return last;
+                        }
+                    }
+                    while (e.MoveNext());
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TSource? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    Func<TSource, bool> predicate = _predicate;
+
+                    foreach (TSource item in _source)
+                    {
+                        if (predicate(item))
+                        {
+                            if (index == 0)
+                            {
+                                found = true;
+                                return item;
+                            }
+
+                            index--;
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public IPartition<TSource>? Skip(int count) => new EnumerablePartition<TSource>(this, count, -1);
+
+            public IPartition<TSource>? Take(int count) => new EnumerablePartition<TSource>(this, 0, count - 1);
         }
 
-        internal sealed partial class WhereArrayIterator<TSource> : IIListProvider<TSource>
+        internal sealed partial class WhereArrayIterator<TSource> : IPartition<TSource>
         {
             public int GetCount(bool onlyIfCheap) => GetCount(onlyIfCheap, _source, _predicate);
 
@@ -118,12 +199,13 @@ namespace System.Linq
                 return result;
             }
 
-            public List<TSource> ToList()
+            public List<TSource> ToList() => ToList(_source, _predicate);
+
+            public static List<TSource> ToList(ReadOnlySpan<TSource> source, Func<TSource, bool> predicate)
             {
                 var list = new List<TSource>();
 
-                Func<TSource, bool> predicate = _predicate;
-                foreach (TSource item in _source)
+                foreach (TSource item in source)
                 {
                     if (predicate(item))
                     {
@@ -133,32 +215,146 @@ namespace System.Linq
 
                 return list;
             }
+
+            public TSource? TryGetFirst(out bool found)
+            {
+                Func<TSource, bool> predicate = _predicate;
+
+                foreach (TSource item in _source)
+                {
+                    if (predicate(item))
+                    {
+                        found = true;
+                        return item;
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TSource? TryGetLast(out bool found)
+            {
+                TSource[] source = _source;
+                Func<TSource, bool> predicate = _predicate;
+
+                for (int i = source.Length - 1; i >= 0; i--)
+                {
+                    if (predicate(source[i]))
+                    {
+                        found = true;
+                        return source[i];
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TSource? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    Func<TSource, bool> predicate = _predicate;
+
+                    foreach (TSource item in _source)
+                    {
+                        if (predicate(item))
+                        {
+                            if (index == 0)
+                            {
+                                found = true;
+                                return item;
+                            }
+
+                            index--;
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public IPartition<TSource>? Skip(int count) => new EnumerablePartition<TSource>(this, count, -1);
+
+            public IPartition<TSource>? Take(int count) => new EnumerablePartition<TSource>(this, 0, count - 1);
         }
 
-        private sealed partial class WhereListIterator<TSource> : Iterator<TSource>, IIListProvider<TSource>
+        private sealed partial class WhereListIterator<TSource> : Iterator<TSource>, IPartition<TSource>
         {
             public int GetCount(bool onlyIfCheap) => WhereArrayIterator<TSource>.GetCount(onlyIfCheap, CollectionsMarshal.AsSpan(_source), _predicate);
 
             public TSource[] ToArray() => WhereArrayIterator<TSource>.ToArray(CollectionsMarshal.AsSpan(_source), _predicate);
 
-            public List<TSource> ToList()
-            {
-                var list = new List<TSource>();
+            public List<TSource> ToList() => WhereArrayIterator<TSource>.ToList(CollectionsMarshal.AsSpan(_source), _predicate);
 
+            public TSource? TryGetFirst(out bool found)
+            {
                 Func<TSource, bool> predicate = _predicate;
+
                 foreach (TSource item in CollectionsMarshal.AsSpan(_source))
                 {
                     if (predicate(item))
                     {
-                        list.Add(item);
+                        found = true;
+                        return item;
                     }
                 }
 
-                return list;
+                found = false;
+                return default;
             }
+
+            public TSource? TryGetLast(out bool found)
+            {
+                ReadOnlySpan<TSource> source = CollectionsMarshal.AsSpan(_source);
+                Func<TSource, bool> predicate = _predicate;
+
+                for (int i = source.Length - 1; i >= 0; i--)
+                {
+                    if (predicate(source[i]))
+                    {
+                        found = true;
+                        return source[i];
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TSource? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    Func<TSource, bool> predicate = _predicate;
+
+                    foreach (TSource item in CollectionsMarshal.AsSpan(_source))
+                    {
+                        if (predicate(item))
+                        {
+                            if (index == 0)
+                            {
+                                found = true;
+                                return item;
+                            }
+
+                            index--;
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public IPartition<TSource>? Skip(int count) => new EnumerablePartition<TSource>(this, count, -1);
+
+            public IPartition<TSource>? Take(int count) => new EnumerablePartition<TSource>(this, 0, count - 1);
         }
 
-        private sealed partial class WhereSelectArrayIterator<TSource, TResult> : IIListProvider<TResult>
+        private sealed partial class WhereSelectArrayIterator<TSource, TResult> : IPartition<TResult>
         {
             public int GetCount(bool onlyIfCheap) => GetCount(onlyIfCheap, _source, _predicate, _selector);
 
@@ -210,47 +406,107 @@ namespace System.Linq
                 return result;
             }
 
-            public List<TResult> ToList()
+            public List<TResult> ToList() => ToList(_source, _predicate, _selector);
+
+            public static List<TResult> ToList(ReadOnlySpan<TSource> source, Func<TSource, bool> predicate, Func<TSource, TResult> selector)
             {
                 var list = new List<TResult>();
 
-                Func<TSource, bool> predicate = _predicate;
-                foreach (TSource item in _source)
+                foreach (TSource item in source)
                 {
                     if (predicate(item))
                     {
-                        list.Add(_selector(item));
+                        list.Add(selector(item));
                     }
                 }
 
                 return list;
             }
+
+            public TResult? TryGetFirst(out bool found) => TryGetFirst(_source, _predicate, _selector, out found);
+
+            public static TResult? TryGetFirst(ReadOnlySpan<TSource> source, Func<TSource, bool> predicate, Func<TSource, TResult> selector, out bool found)
+            {
+                foreach (TSource item in source)
+                {
+                    if (predicate(item))
+                    {
+                        found = true;
+                        return selector(item);
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TResult? TryGetLast(out bool found) => TryGetLast(_source, _predicate, _selector, out found);
+
+            public static TResult? TryGetLast(ReadOnlySpan<TSource> source, Func<TSource, bool> predicate, Func<TSource, TResult> selector, out bool found)
+            {
+                for (int i = source.Length - 1; i >= 0; i--)
+                {
+                    if (predicate(source[i]))
+                    {
+                        found = true;
+                        return selector(source[i]);
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TResult? TryGetElementAt(int index, out bool found) => TryGetElementAt(_source, _predicate, _selector, index, out found);
+
+            public static TResult? TryGetElementAt(ReadOnlySpan<TSource> source, Func<TSource, bool> predicate, Func<TSource, TResult> selector, int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    foreach (TSource item in source)
+                    {
+                        if (predicate(item))
+                        {
+                            if (index == 0)
+                            {
+                                found = true;
+                                return selector(item);
+                            }
+
+                            index--;
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public IPartition<TResult>? Skip(int count) => new EnumerablePartition<TResult>(this, count, -1);
+
+            public IPartition<TResult>? Take(int count) => new EnumerablePartition<TResult>(this, 0, count - 1);
         }
 
-        private sealed partial class WhereSelectListIterator<TSource, TResult> : IIListProvider<TResult>
+        private sealed partial class WhereSelectListIterator<TSource, TResult> : IPartition<TResult>
         {
             public int GetCount(bool onlyIfCheap) => WhereSelectArrayIterator<TSource, TResult>.GetCount(onlyIfCheap, CollectionsMarshal.AsSpan(_source), _predicate, _selector);
 
             public TResult[] ToArray() => WhereSelectArrayIterator<TSource, TResult>.ToArray(CollectionsMarshal.AsSpan(_source), _predicate, _selector);
 
-            public List<TResult> ToList()
-            {
-                var list = new List<TResult>();
+            public List<TResult> ToList() => WhereSelectArrayIterator<TSource, TResult>.ToList(CollectionsMarshal.AsSpan(_source), _predicate, _selector);
 
-                Func<TSource, bool> predicate = _predicate;
-                foreach (TSource item in CollectionsMarshal.AsSpan(_source))
-                {
-                    if (predicate(item))
-                    {
-                        list.Add(_selector(item));
-                    }
-                }
+            public TResult? TryGetElementAt(int index, out bool found) => WhereSelectArrayIterator<TSource, TResult>.TryGetElementAt(CollectionsMarshal.AsSpan(_source), _predicate, _selector, index, out found);
 
-                return list;
-            }
+            public TResult? TryGetFirst(out bool found) => WhereSelectArrayIterator<TSource, TResult>.TryGetFirst(CollectionsMarshal.AsSpan(_source), _predicate, _selector, out found);
+
+            public TResult? TryGetLast(out bool found) => WhereSelectArrayIterator<TSource, TResult>.TryGetLast(CollectionsMarshal.AsSpan(_source), _predicate, _selector, out found);
+
+            public IPartition<TResult>? Skip(int count) => new EnumerablePartition<TResult>(this, count, -1);
+
+            public IPartition<TResult>? Take(int count) => new EnumerablePartition<TResult>(this, 0, count - 1);
         }
 
-        private sealed partial class WhereSelectEnumerableIterator<TSource, TResult> : IIListProvider<TResult>
+        private sealed partial class WhereSelectEnumerableIterator<TSource, TResult> : IPartition<TResult>
         {
             public int GetCount(bool onlyIfCheap)
             {
@@ -316,6 +572,87 @@ namespace System.Linq
 
                 return list;
             }
+
+            public TResult? TryGetFirst(out bool found)
+            {
+                Func<TSource, bool> predicate = _predicate;
+
+                foreach (TSource item in _source)
+                {
+                    if (predicate(item))
+                    {
+                        found = true;
+                        return _selector(item);
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TResult? TryGetLast(out bool found)
+            {
+                using IEnumerator<TSource> e = _source.GetEnumerator();
+
+                if (e.MoveNext())
+                {
+                    Func<TSource, bool> predicate = _predicate;
+                    TSource? last = default;
+                    do
+                    {
+                        TSource current = e.Current;
+                        if (predicate(current))
+                        {
+                            last = current;
+                            found = true;
+
+                            while (e.MoveNext())
+                            {
+                                current = e.Current;
+                                if (predicate(current))
+                                {
+                                    last = current;
+                                }
+                            }
+
+                            return _selector(last);
+                        }
+                    }
+                    while (e.MoveNext());
+                }
+
+                found = false;
+                return default;
+            }
+
+            public TResult? TryGetElementAt(int index, out bool found)
+            {
+                if (index >= 0)
+                {
+                    Func<TSource, bool> predicate = _predicate;
+
+                    foreach (TSource item in _source)
+                    {
+                        if (predicate(item))
+                        {
+                            if (index == 0)
+                            {
+                                found = true;
+                                return _selector(item);
+                            }
+
+                            index--;
+                        }
+                    }
+                }
+
+                found = false;
+                return default;
+            }
+
+            public IPartition<TResult>? Skip(int count) => new EnumerablePartition<TResult>(this, count, -1);
+
+            public IPartition<TResult>? Take(int count) => new EnumerablePartition<TResult>(this, 0, count - 1);
         }
     }
 }